+++ src/cases/view-model/task-list.model.js
@@ -1,5 +1,5 @@
-import { createCaseFromData } from '../../common/helpers/createCaseFromData.js'
-import { processCaseWithWorkflow } from '../../common/helpers/processCaseWithWorkflow.js'
+import { createCaseFromData } from '../../common/helpers/create-case-from-data.js'
+import { processCaseWithWorkflow } from '../../common/helpers/process-case-with-workflow.js'
 
 export const createTaskListViewModel = async (caseData, workflow) => {
   const caseAndStageData = await processCaseWithWorkflow(
