+++ src/common/helpers/process-case-with-workflow.js
